添加box2d

+++ desktop/src/com/mygdx/game/desktop/DesktopLauncher.java
@@ -7,8 +7,8 @@ import com.mygdx.game.MarioBros;
 public class DesktopLauncher {
 	public static void main (String[] arg) {
 		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
-		config.width = MarioBros.V_WIDTH;
-		config.height = MarioBros.V_HEIGHT;
+		config.width = (int)MarioBros.V_WIDTH;
+		config.height = (int)MarioBros.V_HEIGHT;
 		new LwjglApplication(new MarioBros(), config);
 	}
 }
